에스크로로 으로보내기

+++ src/main/java/com/ddiring/backend_asset/api/escrow/EscrowDto.java
@@ -4,8 +4,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.math.BigDecimal;
-
 
 @Getter
 @Setter
@@ -15,5 +13,5 @@ public class EscrowDto {
     private String userSeq;
     private Integer transSeq;
     private Integer transType;
-    private BigDecimal amount;
+    private Integer amount;
 }
